Import Service fix and Session Logger implementation

+++ lib/src/features/documents/cubit/documents_cubit.dart
@@ -18,7 +18,7 @@ class DocumentsCubit extends Cubit<DocumentsState> with SequentialHandler {
   final DataSource dataSource;
 
   DocumentsCubit({required this.dataSource}) : super(DocumentsState.initial()) {
-    MyClassObserver.instance.onCreate(name);
+    SessionLogger.instance.onCreate(name);
     _loadData();
   }
 
@@ -26,19 +26,19 @@ class DocumentsCubit extends Cubit<DocumentsState> with SequentialHandler {
 
   @override
   void emit(DocumentsState state) {
-    MyClassObserver.instance.onTransition(name, this.state, state);
+    SessionLogger.instance.onTransition(name, this.state, state);
     super.emit(state);
   }
 
   @override
   void onError(Object error, StackTrace stackTrace) {
     super.onError(error, stackTrace);
-    MyClassObserver.instance.onError(name, error, stackTrace);
+    SessionLogger.instance.onError(name, error, stackTrace);
   }
 
   @override
   Future<void> close() {
-    MyClassObserver.instance.onClose(name);
+    SessionLogger.instance.onClose(name);
     return super.close();
   }
 
@@ -68,7 +68,7 @@ class DocumentsCubit extends Cubit<DocumentsState> with SequentialHandler {
             stackTrace: s,
           ),
         );
-        MyClassObserver.instance.onError(name, e, s);
+        SessionLogger.instance.onError(name, e, s);
       } finally {
         emit(DocumentsState.idle(documents: state.documents));
       }
@@ -121,7 +121,7 @@ class DocumentsCubit extends Cubit<DocumentsState> with SequentialHandler {
     },
   );
 
-Future<void> addAllDocuments(List<Document> documents, {bool replace = false}) {
+Future<void> addAllDocuments(List<Document> documents, {bool replace = true}) {
   return handle(() async {
     emit(
       DocumentsState.processing(
@@ -131,19 +131,7 @@ Future<void> addAllDocuments(List<Document> documents, {bool replace = false}) {
     );
 
     try {
-      List<Document> finalDocuments;
-      
-      if (replace) {
-        final allIds = state.documents?.map((e) => e.id).toList() ?? [];
-        if (allIds.isNotEmpty) {
-          await deleteDocuments(allIds);
-        }
-        finalDocuments = documents;
-      } else {
-        finalDocuments = [...?state.documents, ...documents];
-      }
-
-      final insertedDocs = await dataSource.insertAllDocuments(documents);
+      final insertedDocs = await dataSource.insertAllDocuments(documents, replace: replace);
       
       for (final doc in insertedDocs) {
         final currentVersion = doc.versions.firstWhere(
@@ -162,7 +150,7 @@ Future<void> addAllDocuments(List<Document> documents, {bool replace = false}) {
 
       emit(
         DocumentsState.processing(
-          documents: finalDocuments,
+          documents: insertedDocs,
           message: "Documents ${replace ? 'replaced' : 'added'} successfully",
         ),
       );
@@ -175,7 +163,7 @@ Future<void> addAllDocuments(List<Document> documents, {bool replace = false}) {
           stackTrace: s,
         ),
       );
-      MyClassObserver.instance.onError(name, e, s);
+      SessionLogger.instance.onError(name, e, s);
     } finally {
       emit(DocumentsState.idle(documents: state.documents));
     }
@@ -217,7 +205,7 @@ Future<void> addAllDocuments(List<Document> documents, {bool replace = false}) {
             stackTrace: s,
           ),
         );
-        MyClassObserver.instance.onError(name, e, s);
+        SessionLogger.instance.onError(name, e, s);
       } finally {
         emit(DocumentsState.idle(documents: state.documents));
       }
@@ -264,7 +252,7 @@ Future<void> addAllDocuments(List<Document> documents, {bool replace = false}) {
             stackTrace: s,
           ),
         );
-        MyClassObserver.instance.onError(name, e, s);
+        SessionLogger.instance.onError(name, e, s);
       } finally {
         emit(DocumentsState.idle(documents: state.documents));
       }
@@ -301,7 +289,7 @@ Future<void> addAllDocuments(List<Document> documents, {bool replace = false}) {
             stackTrace: s,
           ),
         );
-        MyClassObserver.instance.onError(name, e, s);
+        SessionLogger.instance.onError(name, e, s);
       } finally {
         emit(DocumentsState.idle(documents: state.documents));
       }
@@ -355,7 +343,7 @@ Future<void> addAllDocuments(List<Document> documents, {bool replace = false}) {
             stackTrace: s,
           ),
         );
-        MyClassObserver.instance.onError(name, e, s);
+        SessionLogger.instance.onError(name, e, s);
       } finally {
         emit(DocumentsState.idle(documents: state.documents));
       }
@@ -366,6 +354,7 @@ Future<void> addAllDocuments(List<Document> documents, {bool replace = false}) {
     required String title,
     required String? originalPath,
     bool isFavorite = false,
+    bool replace = true,
     int? folderId,
     String? comment,
     DateTime? expirationDate,
@@ -383,7 +372,7 @@ Future<void> addAllDocuments(List<Document> documents, {bool replace = false}) {
 
     try {
       final safePath = await FileService.saveFileToAppDir(originalPath);
-      await FileService.deleteFile(originalPath);
+      if(replace) await FileService.deleteFile(originalPath);
       final doc = Document(
         id: 0,
         title: title,
@@ -407,7 +396,7 @@ Future<void> addAllDocuments(List<Document> documents, {bool replace = false}) {
 
       return ResultOr.success(safePath);
     } catch (error, stackTrace) {
-      MyClassObserver.instance.onError(
+      SessionLogger.instance.onError(
         name,
         error,
         stackTrace,
@@ -433,7 +422,7 @@ Future<void> addAllDocuments(List<Document> documents, {bool replace = false}) {
       );
       return sizes.fold<int>(0, (sum, size) => sum + size);
     } catch (e, s) {
-      MyClassObserver.instance.onError(
+      SessionLogger.instance.onError(
         name,
         e,
         s,
